fix(rng): retail-width sweep — mock RNG copy, RMG oracle stubs, referee Q_rsqrt
Systemic follow-through on the holdrand width ruling (d43c2bce):
- host-interface mock carried a second platform-width holdrand replica
  (feeds the RMG oracle dumper seam) -> u32.
- tools/rmg-oracle host stubs: same 32-bit patch; seed/dedicated
  goldens regenerated from the retail-width dumper.
- tools/referee-oracle/build.sh: normalize Q_rsqrt's `long i` type-pun
  to int (LP64 reads 4 bytes of stack garbage past the float — UB and
  divergent from the retail i386 32-bit long; jampgame-oracle's dumper
  already normalized it, the referee dylib did not).

Oracle long-arithmetic survey (game/qcommon): remaining surfaces are
bg_lib's QVM-only qsort machinery (retail links CRT qsort) and one
small-valued trap arg — no further gameplay-relevant width leaks.
Referee 9/9, workspace green.

+++ tools/rmg-oracle/src/rmg_host_stubs.cpp
@@ -129,7 +129,7 @@ CCMShader *CM_GetShaderInfo( const char *name ) {
 // Faithful q_math.c helpers (the determinism anchor for golden #1). Transcribed
 // VERBATIM from oracle/codemp/game/q_math.c:1432-1470 and :SetPlaneSignbits.
 // --------------------------------------------------------------------------
-static unsigned long holdrand = 0x89abcdef;
+static unsigned int holdrand = 0x89abcdef; /* retail-win32 32-bit width (2026-07-17 ruling) */
 
 void Rand_Init( int seed ) { holdrand = seed; }
 unsigned long rng_state( void ) { return holdrand; }
